Refactor GeoJson parsing, to be the same with rsp rasterize and rsp cover. Fix #29 and Fix #68
